Revert "refactor(blockchain-link): name worker source files"
This reverts commit 44157bc317d9f99cba4f5427ffa2a7c616672350.
