+++ packages/blockchain-link/jest.config.unit.cjs
@@ -14,13 +14,13 @@ module.exports = {
         'src/types',
         'src/ui',
         'fixtures',
-        'src/workers/blockbook/blockbookWorker.test.ts',
-        'src/workers/blockfrost/blockfrostWorker.test.ts',
+        'src/workers/blockbook/blockbook.test.ts',
+        'src/workers/blockfrost/blockfrost.test.ts',
         'src/workers/connection.test.ts',
-        'src/workers/electrum/electrumWorker.test.ts',
-        'src/workers/ripple/rippleWorker.test.ts',
-        'src/workers/solana/solanaWorker.test.ts',
-        'src/workers/stellar/stellarWorker.test.ts',
+        'src/workers/electrum/electrum.test.ts',
+        'src/workers/ripple/ripple.test.ts',
+        'src/workers/solana/solana.test.ts',
+        'src/workers/stellar/stellar.test.ts',
     ],
     setupFiles: ['./setup.js'],
     testEnvironment: '../../JestCustomEnv.js',
