+++ packages/blockchain-link/src/__fixtures__/allTestWorkers.ts
@@ -1,6 +1,6 @@
-import Blockbook from '../workers/blockbook/blockbookWorker';
-import Blockfrost from '../workers/blockfrost/blockfrostWorker';
-import Ripple from '../workers/ripple/rippleWorker';
+import Blockbook from '../workers/blockbook';
+import Blockfrost from '../workers/blockfrost';
+import Ripple from '../workers/ripple';
 
 export const allTestWorkers = [
     { name: 'blockbook', worker: Blockbook },
