+++ packages/blockchain-link/src/getAccountInfo.test.ts
@@ -2,7 +2,8 @@ import { BackendWebsocketServerMock } from '@trezor/e2e-utils';
 
 import { allTestWorkers } from './__fixtures__/allTestWorkers';
 import fixtures from './__fixtures__/getAccountInfo';
-import { BlockchainLink } from './blockchainLink';
+
+import { BlockchainLink } from './index';
 
 allTestWorkers.forEach(instance => {
     describe(`getAccountInfo: ${instance.name}`, () => {
