+++ packages/blockchain-link/src/subscribe.test.ts
@@ -2,7 +2,8 @@ import { BackendWebsocketServerMock } from '@trezor/e2e-utils';
 
 import { allTestWorkers } from './__fixtures__/allTestWorkers';
 import fixtures from './__fixtures__/subscribe';
-import { BlockchainLink } from './blockchainLink';
+
+import { BlockchainLink } from './index';
 
 // this test covers application live cycle
 // where "subscribe" and "unsubscribe" is called multiple times on single blockchain-link instance
