+++ packages/blockchain-link/src/workers/blockbook/blockbook.test.ts
@@ -1,32 +1,32 @@
 import { BackendWebsocketServerMock } from '@trezor/e2e-utils';
 
-import { BlockchainLink } from '../../blockchainLink';
+import { BlockchainLink } from '../../index';
 import {
-    //  blockfrostWorkerFactory,
-    blockfrostModuleFactory,
+    // blockbookWorkerFactory,
+    blockbookModuleFactory,
 } from '../__fixtures__/integrationWorker';
 
 const backends = [
     // TODO: nodejs tests are failing, fix it.
     // {
-    //     name: 'nodejs-build:blockfrost',
-    //     type: 'blockfrost',
-    //     worker: blockfrostWorkerFactory,
+    //     name: 'nodejs-build:blockbook',
+    //     type: 'blockbook',
+    //     worker: blockbookWorkerFactory,
     // },
     {
-        name: 'module-build:blockfrost',
-        type: 'blockfrost',
-        worker: blockfrostModuleFactory,
+        name: 'module-build:blockbook',
+        type: 'blockbook',
+        worker: blockbookModuleFactory,
     },
 ];
 
 backends.forEach(b => {
-    describe(`Blockfrost ${b.name}`, () => {
+    describe(`Blockbook ${b.name}`, () => {
         let server: BackendWebsocketServerMock;
         let blockchain: BlockchainLink;
 
         beforeEach(async () => {
-            server = await BackendWebsocketServerMock.create('blockfrost');
+            server = await BackendWebsocketServerMock.create('blockbook');
             blockchain = new BlockchainLink({
                 name: b.name,
                 worker: b.worker,
@@ -43,15 +43,12 @@ backends.forEach(b => {
         it('Get info', async () => {
             const result = await blockchain.getInfo();
             expect(result).toEqual({
-                name: 'BlockfrostMock',
-                shortcut: 'ada',
-                network: 'ada',
-                decimals: 6,
+                name: 'TestMock',
+                shortcut: 'test',
+                network: 'test',
+                decimals: 9,
                 blockHeight: 1,
-                blockHash: 'test_block_hash-hash',
-                testnet: false,
                 url: expect.any(String),
-                version: '1.4.0',
             });
         });
