+++ packages/blockchain-link/src/workers/ripple/ripple.test.ts
@@ -1,6 +1,6 @@
 import { BackendWebsocketServerMock } from '@trezor/e2e-utils';
 
-import { BlockchainLink } from '../../blockchainLink';
+import { BlockchainLink } from '../../index';
 import {
     // rippleWorkerFactory,
     rippleModuleFactory,
