+++ packages/blockchain-link/src/workers/stellar/stellar.test.ts
@@ -3,8 +3,9 @@ import * as utils from '@trezor/blockchain-link-utils/src/stellar';
 import { getStellarConnection, identifyTransaction, toStroops } from '@trezor/network-stellar';
 import type { StellarAPI } from '@trezor/network-stellar/types';
 
-import StellarWorker from './stellarWorker';
-import { BlockchainLink } from '../../blockchainLink';
+import { BlockchainLink } from '../../index';
+
+import StellarWorker from './index';
 
 const HORIZON_URL = 'https://horizon.stellar.org';
